requirements.txt
discord.py
irc
google-genai